+++ src/lib.rs
@@ -111,6 +111,7 @@ detected.
   * [Scalable Vector Graphics (SVG)](`FileFormat::ScalableVectorGraphics`)
   * [Simple Object Access Protocol (SOAP)](`FileFormat::SimpleObjectAccessProtocol`)
   * [Training Center XML (TCX)](`FileFormat::TrainingCenterXml`)
+  * [Universal Subtitle Format (USF)](`FileFormat::UniversalSubtitleFormat`)
   * [XML Localization Interchange File Format (XLIFF)](`FileFormat::XmlLocalizationInterchangeFileFormat`)
   * [XML Shareable Playlist Format (XSPF)](`FileFormat::XmlShareablePlaylistFormat`)
   * [draw.io (DRAWIO)](`FileFormat::Drawio`)